WIP Put attachments in different files.
