+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/OmsorgEgenAuto.kt
@@ -1,10 +1,9 @@
 package no.nav.pensjon.brev.maler
 
-import no.nav.pensjon.brev.api.model.Felles
 import no.nav.pensjon.brev.api.model.LetterMetadata
-import no.nav.pensjon.brev.api.model.NAVEnhet
-import no.nav.pensjon.brev.api.model.ReturAdresse
 import no.nav.pensjon.brev.api.model.maler.OmsorgEgenAutoDto
+import no.nav.pensjon.brev.maler.vedlegg.EgenerklaeringPleieOgOmsorgsarbeid
+import no.nav.pensjon.brev.maler.vedlegg.egenerklaeringPleieOgOmsorgsarbeid
 import no.nav.pensjon.brev.template.*
 import no.nav.pensjon.brev.template.Language.*
 import no.nav.pensjon.brev.template.base.PensjonLatex
@@ -28,148 +27,42 @@ object OmsorgEgenAuto : StaticTemplate {
         )
     ) {
 
-        val aarEgenerklaring = argument().select(OmsorgEgenAutoDto::aarEgenerklaringOmsorgspoeng).str()
+        val aarEgenerklaringOmsorgspoeng = argument().select(OmsorgEgenAutoDto::aarEgenerklaringOmsorgspoeng).str()
 
         outline {
             paragraph {
                 textExpr(
                     Bokmal to
                             "Vi trenger en bekreftelse på at du har utført pleie- og omsorgsarbeid i ".expr()
-                            + aarEgenerklaring
+                            + aarEgenerklaringOmsorgspoeng
                             + ". Derfor må du fylle ut det vedlagte skjemaet og sende det til oss innen fire uker.",
 
                     Nynorsk to
                             "Vi treng ei stadfesting på at du har utført pleie- og omsorgsarbeid i ".expr()
-                            + aarEgenerklaring
+                            + aarEgenerklaringOmsorgspoeng
                             + ". Du må difor nytte det vedlagde skjemaet og sende til oss innan fire veker.",
 
                     English to
                             "We need you to confirm that you have provided nursing and care work in ".expr()
-                            + aarEgenerklaring
+                            + aarEgenerklaringOmsorgspoeng
                             + ". Therefore, it is required that you complete the enclosed form and return it to NAV within four weeks."
                 )
             }
 
             paragraph {
-                val aarInnvilget = argument().select(OmsorgEgenAutoDto::aarInnvilgetOmsorgspoeng).str()
+                val aarInnvilgetOmsorgspoeng = argument().select(OmsorgEgenAutoDto::aarInnvilgetOmsorgspoeng).str()
                 textExpr(
-                    Bokmal to "Du har fått godkjent pensjonsopptjening for ".expr() + aarInnvilget + ".",
-                    Nynorsk to "Du har fått godkjend pensjonsopptening for ".expr() + aarInnvilget + ".",
-                    English to "You have accumulated pensionable earnings for ".expr() + aarInnvilget + "."
+                    Bokmal to "Du har fått godkjent pensjonsopptjening for ".expr() + aarInnvilgetOmsorgspoeng + ".",
+                    Nynorsk to "Du har fått godkjend pensjonsopptening for ".expr() + aarInnvilgetOmsorgspoeng + ".",
+                    English to "You have accumulated pensionable earnings for ".expr() + aarInnvilgetOmsorgspoeng + "."
                 )
             }
 
         }
 
-        attachment(
-            title = newText(
-                Bokmal to "Egenerklæring om pleie- og omsorgsarbeid",
-                Nynorsk to "Eigenmelding om pleie- og omsorgsarbeid",
-                English to "Personal declaration that nursing and care work has been provided",
-            ),
-            includeSakspart = true,
-        ) {
-            paragraph {
-                val dokDato = felles().select(Felles::dokumentDato).format()
-                textExpr(
-                    Bokmal to "Jeg viser til brev av ".expr() + dokDato + ".",
-                    Nynorsk to "Eg viser til brev datert ".expr() + dokDato + ".",
-                    English to "I refer to your letter dated ".expr() + dokDato + ".",
-                )
-            }
-
-            paragraph {
-                textExpr(
-                    Bokmal to "I ".expr() + aarEgenerklaring + " har jeg utført pleie og omsorgsarbeid på minst 22 timer i uken. (Inkludert opptil en halv time reisetid per besøk.)",
-                    Nynorsk to "I ".expr() + aarEgenerklaring + " har eg utført pleie- og omsorgsarbeid på minst 22 timar i veka. (Inkludert opptil ein halv time reisetid per besøk.)",
-                    English to "In ".expr() + aarEgenerklaring + " I have provided care work that has amounted to at least 22 hours per week. (Travelling time up to 30 minutes per visit may be included.)",
-                )
-            }
-
-            formText(
-                60, newText(
-                    Bokmal to "Navn på pleietrengende:",
-                    Nynorsk to "Navn på pleietrengende:",
-                    English to "I have provided care work for:",
-                )
-            )
-
-            formChoice(
-                newText(
-                    Bokmal to "Arbeidet har vart i (sett kryss):",
-                    Nynorsk to "Arbeidet har vart i (set kryss):",
-                    English to "The work has lasted for (insert X):"
-                )
-            ) {
-                choice(
-                    Bokmal to "minst seks måneder",
-                    Nynorsk to "minst seks månader",
-                    English to "at least six months"
-                )
-                choice(
-                    Bokmal to "under seks måneder",
-                    Nynorsk to "under seks månader",
-                    English to "less than six months"
-                )
-            }
-
-            formText(
-                size = 0, prompt = newText(
-                    Bokmal to "Hvis omsorgsforholdet har opphørt i løpet av året:",
-                    Nynorsk to "Om omsorgsforholdet er blitt avslutta under året:",
-                    English to "If care work has ceased during the year:",
-                )
-            )
-            formText(
-                size = 25,
-                vspace = false,
-                prompt = newText(
-                    Bokmal to "Oppgi dato for opphøret:",
-                    Nynorsk to "Dato for opphøyr:",
-                    English to "State date if ceased:"
-                )
-            )
-            formText(
-                size = 55,
-                vspace = false,
-                prompt = newText(
-                    Bokmal to "Oppgi årsaken til opphøret:",
-                    Nynorsk to "Grunnen til opphøyr: ",
-                    English to "State reason if ceased"
-                )
-            )
-
-            repeat(4) { newline() }
-
-            formText(size = 25, prompt = newText(Bokmal to "Dato:", Nynorsk to "Dato:", English to "Date"))
-            formText(
-                size = 55,
-                vspace = false,
-                prompt = newText(Bokmal to "Underskrift:", Nynorsk to "Underskrift:", English to "Signature:")
-            )
-
-            repeat(3) { newline() }
-
-            paragraph {
-                text(
-                    Bokmal to "Du må sende denne egenerklæringen til:",
-                    Nynorsk to "Du må sende denne eigenmeldinga til:",
-                    English to "Please return the form to:"
-                )
-                newline()
-
-                val avsender = felles().select(Felles::avsenderEnhet)
-                eval { avsender.select(NAVEnhet::navn) }
-                newline()
-
-                val returAdresse = avsender.select(NAVEnhet::returAdresse)
-                eval { returAdresse.select(ReturAdresse::adresseLinje1) }
-                newline()
-
-                eval(returAdresse.select(ReturAdresse::postNr) + " " + returAdresse.select(ReturAdresse::postSted))
-            }
-        }
-
+        includeAttachment(egenerklaeringPleieOgOmsorgsarbeid, argument().map {
+            EgenerklaeringPleieOgOmsorgsarbeid(it.aarEgenerklaringOmsorgspoeng)
+        })
     }
 
 }
